% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{clean_categorical}
\alias{clean_categorical}
\title{Clean Categorical Variables}
\usage{
clean_categorical(df, valid_values)
}
\arguments{
\item{df}{A data frame.}

\item{valid_values}{A list where each element is a vector of valid values for a categorical column.}
}
\value{
A cleaned data frame with only valid categorical values.
}
\description{
This function replaces invalid categorical values with mode of the column.
}
