% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{clean_continuous}
\alias{clean_continuous}
\title{Clean Continuous Variables}
\usage{
clean_continuous(df, cols)
}
\arguments{
\item{df}{A data frame containing numerical variables.}

\item{cols}{A vector of column names to clean.}
}
\value{
A cleaned data frame with outliers replaced by column means
}
\description{
This function removes outliers using the interquartile range (IQR) method and ensures integer-only columns are rounded.
}
