% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{clean_data}
\alias{clean_data}
\title{Clean Raw Clinical Dataset}
\usage{
clean_data(file_path, preprocess = TRUE)
}
\arguments{
\item{file_path}{Path to the raw dataset file.}
}
\value{
A cleaned dataframe with selected columns.
}
\description{
Reads, processes, and cleans a clinical dataset file.
}
